Modify test as a result of credit mining refactor
